Move Launcher to a base module, like providers/executors/channels
This PR moves the base class to a *.base module, instead of defining it
it in parsl.launchers.launchers.

This makes the interface for users more consistent with providers,
executors and channels, with the base class provided in *.base, and
common implementations exported from the top level module.

This is a breaking change for anyone who imports Launcher,
for example:
* anyone who has implemented configuration typing outside of parsl
* anyone who has implemented their own Launcher)
